rename cli folder to cmd to follow Go convention
