+++ cmd/bblfsh-sdk/cmd/build.go
@@ -4,7 +4,7 @@ import (
 	"path/filepath"
 
 	"gopkg.in/bblfsh/sdk.v1/assets/build"
-	"gopkg.in/bblfsh/sdk.v1/cli"
+	"gopkg.in/bblfsh/sdk.v1/cmd"
 )
 
 const sdkPath = ".sdk"
@@ -12,7 +12,7 @@ const sdkPath = ".sdk"
 const PrepareBuildCommandDescription = "installs locally the build system for a driver"
 
 type PrepareBuildCommand struct {
-	cli.Command
+	cmd.Command
 }
 
 func (c *PrepareBuildCommand) Execute(args []string) error {
